Changing some installation paths.
"broctl install" copied a number of files to share/bro/*, which
violates the common assumption that things there are static. It can
also create permission problems if the user running "broctl install"
is not the one installing Bro. So now the pieces copied/generated by
"broctl install" are moved to spool/*.

Specifically, that affects share/bro/.site and
share/broctl/scripts/broctl-config.sh. The former is now in
spool/policy/{site,auto} and the latter in spool/broctl-config.sh.
share/broctl/scripts/broctl-config.sh is however made a link to
point to spool/.

This however means that even in NFS-mode, we need to rsync the stuff
now in spool/.

More changes:

     - The CMake install does no longer recreate some of the
     top-level directories when they already exist. That makes it
     possible to now symlink them somewhere else after the first
     install.

     - When broctl doesn't find spool/broctl.dat it no longer
     aborts but just warns. That allows CMake to skip installing an
     empty one.

     - Deleting an unused policy file.