Fixing process command.
After the script reorg, it was now activing log rotation. I've moved
that part out of manager.bro/standalone.bro into enable-rotation.bro,
which is pulled in only in live operation.

Seth, you should probably check that this is right, and that the
policy files are located where they belong.