Fix InstallShellScript.cmake module to recognize changes to original source.
Shell scripts are now configure_file()'d into the build directory,
then hashbang-normalized (setting '#!' to use full path to shell
interpreter), and then the normalized version is scheduled for
installation.

Previously, the configure_file() on original source wasn't done and
instead the hashbang-normalized version was configured in-place, which meant
changes to the original source code wouldn't trigger the internal make
dependencies in order to regenerate the configured/normalized file.