Fix MailAlarmsTo broctl config option
This change writes the value of the broctl config option MailAlarmsTo
to the appropriate bro script variable, so that alarm summary emails
are sent to the address specified with this config option (previously,
this config option was not used anywhere).  As before, if the user does not
specify a value for this option, then the value of the MailTo config
option is used by default.

Addresses #814.