Fix `broctl install` to now also copy subdirs of dirs in SitePolicyPath.
Paths in SitePolicyPath were expected to be a flat directory of bro
scripts, so adding a new path to the default SitePolicyPath and then
adding bro script "module" directories to it would not be correctly
installed by BroControl.  Instead of silently ignoring subdirs
inside a path in SitePolicyPath, the install command now recursively
copies them to the right place.

Addresses #902.