Fix various bugs and remove some unused code
Removed an unused extra parameter from the cmd_scripts_pre and
cmd_scripts_post methods (this was causing the TestPlugin.py to
crash broctl when running the "scripts" command).

Removed an undefined "cleanup" command parameter "--keep-tmp" that the
"restart --clean" command was trying to use.

The "status" command was not calling cmd_status_post (it was calling
cmd_status_pre twice).

Removed the unused cmd_restart_pre and cmd_restart_post methods.

Removed the unused function "completedefault" (it was using an
incorrect list of command names).