Add return value to some cmd_<cmd>_pre methods
Fixed the cmd_<cmd>_pre methods in plugin API that don't take a node
argument to return a boolean indicating whether or not the
corresponding <cmd> should run (as already documented).
Updated documentation for the affected methods to make this more clear.
Changed broctl to check for these return values.

Updated the TestPlugin to return True for the affected methods.