Changed return value of plugin API "execute" method
Changed the return value of the plugin API "execute" method to
not include the node (to match the documented behavior).

Also changed some helper scripts so that the return value is
sent after the script output (instead of before), so that we can
get the actual exit code, and so that the "success" boolean (in
"execute" method return value) has the correct value.