Fix broctl plugin option names to be case-insensitive
Fixed the handling of plugin option names so that they are stored in
lowercase in order to avoid duplicate entries (differing only
by case) if someone defines an option in a plugin that is not all
lowercase (because options in broctl.cfg were already being converted
to lowercase).  Changed the _setOption method to do this automatically
so we don't have to remember to do this everywhere it is called.