Fix plugin inconsistency for certain broctl commands
The plugin cmd_<cmd>_pre method was receiving an empty node list
for the start, stop, and restart commands when the user did not
specify any nodes on the command-line.  However, for other
commands, the plugin receives the complete node list when the
user doesn't specify any nodes on the command-line.

To fix this inconsistency, the "expand_all" argument to the
nodeArgs() helper function was removed, and the control.start()
and control.stop() functions were modified to sort out the supplied
list of nodes into each of the three types (manager, proxy, worker)
so that the nodes can be started or stopped in the correct order.

This change also affects the case where a user explicitly supplies
node arguments (such as "stop manager proxies"), in which case the
code will now re-order the supplied nodes instead of just processing
them in the order that the user specified.