Prevent capstats from being run with invalid args
When running broctl cron, capstats is called for all nodes (but only
once for each unique address/interface pair).  However, nodes
that don't have an interface in the broctl configuration would result
in an error message from capstats that was visible in the stats.log
file (because the interface name is missing).  The fix is to skip trying
to run capstats on nodes that don't have a network interface configured.