Add support for multiple PF_RING cluster IDs
Instead of assigning the same PF_RING cluster ID to every worker
in a Bro cluster, the pf_ring broctl plugin has been modified to
automatically assign a different PF_RING cluster ID for each set
of workers on a host that all sniff the same interface.  The first
such set of workers on a host are assigned the globally-configured
PF_RING cluster ID (this is the "pfringclusterid" broctl option in
broctl.cfg).  Each subsequent set of workers on a host that sniff
another interface are assigned a different value (incremented by
one from previous value).

Addresses #943.