Fix a couple of bugs in the broctl ps plugin
Attempting to use the broctl ps plugin always results in a python
stack trace on OS X 10.8, due to a harmless message on stderr when
the "ps" command is run: "dyld: DYLD_ environment variables being
ignored because main executable (/bin/ps) is setuid or setgid".
Fixed by searching the output for the start of the "ps" command output.

Attempting to use the ps plugin (on any OS) also results in a python
stack trace when connecting to a remote node fails.  This is due to
the output being the python None object (instead of a list of strings).
Fixed by checking the output before attempting to access it.