Fix a bug when the time command is not found
When the time command was not found, broctl was setting an error
message as the value of the broctl "time" option (this would result in
connection summary reports not being generated).  If there was no
error message, then broctl would immediately exit with a python stack
trace.  Fixed by checking the return value of the captureCmd method, and
setting the "time" option to an empty string (this ensures that the
connection summary reports can still be generated) upon failure.