Allow multiple conn-summary.log files to be processed
If a broctl stop (or restart) happens just after a normal log
rotation occurs, then it is likely that the first pair of
conn.log/conn-summary.log is still being archived while the second
pair is beginning to be processed.  For the conn.log, this poses
no problems, because Bro rotates logs with a name containing a timestamp.
But the conn-summary.log was being created (before it gets archived)
without a timestamp in the filename.  Therefore, two processes were
trying to create two different files in the same directory at the same
time with the same filename.  Fixed by renaming the conn-summary.log
to contain a timestamp.