Fix a race condition during broctl start
During a "broctl start" one process creates a PID file, while
a second process waits for the file to be created and then reads
the contents of the file.  If the second process reads the file
after it has been opened but before the contents are written, then
broctl was being terminated with a python stack trace.  Fixed by
making the second process wait until the file exists and has
nonzero size.