Fix broctl exec command to check for errors
Fixed a bug where broctl was being terminated with a python stack trace
if the "exec" command was run and we couldn't connect to a remote host
(this was due to the executeCmdsParallel function returning None for
the output return value). Now it will correctly report an error.

Also updated the plugin API documentation to mention this issue so that
similar errors can be avoided when writing a plugin.