Fix usage of PF_RING interface containing semicolons
When using PF_RING and specifying an interface name containing semicolons
(for example, "p2p1;p2p2", this is used to aggregate traffic from
multiple devices), there are two places in broctl where the interface
name needs to be quoted in order to prevent the shell from interpreting
the semicolons as command separators.  The semicolons were causing
broctl cron to hang whenever it ran capstats because the interface was
being specified as the first argument to capstats, and the semicolon
prevented capstats from receiving the other arguments.  Similarly,
when starting bro, the semicolons would prevent bro from receiving
all of the intended command-line arguments.

Also improved the error message when capstats fails in order to give
more information to the user.