Don't import readline
On some systems (seems to be RedHat-based systems, such as RHEL,
CentOS, or Fedora), doing "import readline" causes an eight-byte
escape sequence to be sent to stdout.  This is not user-visible when
running broctl interactively, but causes problems when stdout is
redirected to a file.  As a result, garbage output appears in the
generated HTML broctl docs, and some broctl tests fail.

Fixed by removing the import.  Doing so doesn't appear to result
in any loss of functionality.