Add PID file management to archive-log script
Added PID file management to the archive-log script.  This is
used by the post-terminate script to monitor the status of
the archive-log processes while post-terminate is waiting to
remove the tmp directory.

Removed the arbitrary 10 minute timeout from the post-terminate
script.  Instead, now post-terminate will wait as long as
there is at least one instance of archive-log running.

Addresses #970.