Add support for CPU pinning
Added support for CPU pinning.  To use CPU pinning, a new per-node
option "pin_cpus" must be specified in node.cfg, and the OS must be
either Linux or FreeBSD (if such a node.cfg is used on another OS,
then the "pin_cpus" option is ignored).  CPU pinning only applies
to Bro itself, not to other processes that might be executed (such
as helper scripts).  If the user specifies an invalid CPU number,
then the affected Bro processes will not start, and an error message
will be visible with the broctl diag command.

Addresses #996.