Disabling the "set -x" output during broctl-test-setup.
It's pretty verbose and usually, I believe, not that helpful. It turns
"-x" on at the end, i.e., it's now always on after the script ran,
which I think is fine. That also means that the "bash -x" in
individual tests is unecessary now, though I haven't removed them.

However, if this is loosing too much information, I'd be fine
reverting.