Giving the broctl test suite its own build directory.
This seems to be the better approach: by keeping the standard build
directory separate from the version used for testing, we can avoid
having to switch back and forth with configure. This also means that
"make test" will often not need to rebuild all of Bro (or can do so
incrementally just for changes). With that, I've removed the "make
quick" target again.