Minor cleanup of the build script
Reorganized a few things in the build script.  Renamed a variable for
clarity.  Removed an unused variable.  Removed a redundant mkdir.
Improved an error message.  Now that BROCTL_TEST directory is no
longer hard-coded to be in /tmp, there's not really any reason for
a user to specify their own location.