Add ability to set environment variables in node.cfg
Added the ability to specify environment variables in node.cfg
to be set when Bro is started (these can override values set by
the load-balancing plugins).  This uses the existing "env_vars"
node attribute, but now allows a user to specify it in the node.cfg file.

One limitation is that the value of any such environment variable must
be a constant.  If a value can be determined only at run-time, then a user
would need to write their own plugin to set such a variable (similar to
the load-balancing plugins).

Addresses #1010.