Change implementation of env_vars attribute
Now env_vars is stored internally as a dictionary, and the env_vars
configure option in node.cfg is now changed to a comma-separated
list.  As a result, there is now more error-checking of the
user-supplied string in node.cfg, and there is no longer
any possibility for duplicate environment variables to appear
in debug.log or in the "broctl nodes" output (which should reduce
confusion).