Document and fix the behavior of SitePolicyPath
Changed the behavior of the SitePolicyPath broctl option to not clobber
existing files/directories when copying, in order to match the expected
behavior (directories earlier in the list take precedence over directories
later in the list when duplicate filenames are encountered).  This
change affects only the case where there is more than one directory in
the SitePolicyPath list and duplicate filenames are encountered.

Improved the documentation of the SitePolicyPath option to explain how
it works.

Addresses BIT-714.