Improve CPU pinning documentation and error message
Added a test of the CPU pinning command immediately before attempting to
run Bro with CPU pinning.  Upon failure, a more helpful error message is
generated (visible in the output of the broctl "diag" command).

Also improved the documentation of the pin_cpus option with a specific
example to hopefully avoid confusion about what is a valid CPU number.

Addresses BIT-1068