Fix usage of the broargs broctl option
Fixed a problem that would occur when a command-line argument in broargs
contained a space (and the user quoted the argument correctly in broctl.cfg).
The problem was that the shell scripts that run bro were splitting the
argument into separate arguments (due to the embedded space character).

This example broctl.cfg should now work correctly:
broargs = --filter 'not ip6'