Fix a deadlock when capturing output from local command
When broctl captures output from a command run locally via the
runLocalCmd method, and if the command generates enough output to fill
the OS pipe buffer, then broctl hangs waiting for the child process
to terminate, which never happens because the child process is waiting
for broctl to read from the pipe.  Fixed by using the "communicate"
method (instead of the "wait" method) of the Python subprocess module.