Fix a race condition when sendmail option is empty string
If the broctl sendmail option is set to an empty string (this
indicates that the user does not want email sent), then when
broctl execs the shell script to send mail it just exits immediately
(as expected) which triggers a race condition on some systems that
terminates broctl because the child process exits before broctl
can send input via the stdin pipe.  Fixed by not executing the shell
script when the sendmail option is an empty string.