Fix the broctl "top" command output on Linux
On Linux, the broctl "top" command output was sometimes showing the
wrong suffix for the memory statistics (example: it would show 70K
when the correct value is 70M), because the Linux top command does
not use any suffix when the value is kilobytes (e.g., 70 megabytes is
sometimes shown as 70000).  The solution is to set "def_factor=1024"
when running the "to-bytes.awk" script, but this would also multiply
the PID and CPU usage numbers, so new options were added to specify
the starting and ending fields to apply the conversion.

Also added filtering of the header output of the FreeBSD top command,
in order to avoid processing by the "to-bytes.awk" script (some lines
in the header start with a number).