Fix lookup of nodes
The nodes() method returns a list of nodes which match a given
node name or node type, but was incorrectly returning an empty list in
three cases:  nodes("manager") in a cluster config when the manager
node is not named "manager", nodes("standalone") in a standalone config
if there is no broctl option named "standalone", or nodes(<name>)
where <name> is the name of a node that matches the name of a
broctl option.

Fixed by adding checks for the node types "standalone" and "manager",
and by removing code that was searching the list of broctl option
names (we do not store node names or types there).

Also added a test case.

Addresses BIT-1091