Do not check if the local host is "alive"
The broctl cron and install commands check whether a host is "alive"
or not, but there is no need to do this check for the local host.
Checking the local host results in attempting to ssh to the local host,
which may (depending on ssh config and whether the graphical ssh-askpass
can run successfully) result in the user being prompted by ssh-askpass
to enter "yes" or "no" (for the host key verification).  This can happen
regardless of whether the user is running broctl interactively or just
running the broctl tests.