Add another warning message when a host is not alive
When a host is found to be not alive, a warning message is output (as
expected), but the warning is shown only the first time the check fails.
This is especially problematic for the broctl install command, because
it does not show any indication of a problem after the first such
unsuccessful run.

Added another warning message (worded slightly differently) to avoid
confusion in these situations.  For example, a user ran "broctl install"
twice, but only saw a warning the first time.  Next, running
"broctl start" resulted in a misleading error message that had nothing
to do with the actual problem.