Improve the check-pid helper script
Many of the broctl commands rely on the check-pid helper script, but it
was unreliable on OS X (most, if not all, of the recent test failures
seem to be caused by this) due to an excessive number of child
processes spawned in parallel (10 for each check-pid, for each Bro node).
As a result, occasionally the script would fail to identify a running
Bro process (it would return 1 instead of the expected value of 0).
Fixed by simplifying the script (reduced the number of child processes
by half), and it seems to be much more reliable.