Add ability to convert DER files to Bro script redef
Modified a script from Justin Azoff so that broctl can convert DER
certificate files into a Bro script that redefs the SSL::root_certs
variable.  There is a new broctl option which specifies a directory
that contains the DER files (the files must have file extension of ".der").
The DER files are read (and the new Bro script is created) when the
user runs "broctl install".