Improve testing of broctl start and stop commands
Added several new tests (slow start of Bro, slow stop of Bro, starting
and stopping a large number of nodes on one host, restart with "--clean"
option).  Improved most of the existing tests by testing additional
aspects of the start and stop commands (such as sending crash reports,
logging to stats.log, and more test cases when a node crashes).

The "bro" test script now recognizes two new options in its config file:
"slowstart", and "slowstop".  These can be used to simulate a Bro
that is slow to initialize or that is slow to stop (i.e., cannot
be terminated with SIGTERM, so broctl must fallback to using SIGKILL).

Added a "sendmail" test script that can be used in place of the real
sendmail (the test script is faster and easier to test with).

Added a new function "replaceprefix" that any test can use.  It
replaces the text "@PREFIX@" in a test-specific file with the
test-specific Bro install prefix.

Added a new command-line option "--peers" to the diff-status-output
canonifier that indicates whether we want to include the "Peers" column
of the broctl status output in the diff (when we're not running the real
Bro, then the values in that column are not predictable or useful).