Add broctl option to set PF_RING cluster type
Added a new broctl option "PFRINGClusterType", which allows a user
to specify a PF_RING cluster type (the default is the same cluster
type used in previous versions of broctl).  The broctl pf_ring plugin
will use this information to set the correct environment variable
which is needed by the PF_RING-aware libpcap to set the cluster type.

Addresses BIT-1108