Fix plugin init return values
The myricom plugin wasn't explicitly returning a value from its init
method, so it was being disabled as a result.  Fixed return values
in the init method in the myricom and pf_ring plugins.  Also added a
debug message when a plugin is disabled so that this issue is easier
to diagnose in the future.