Improve broctl output formatting
Increased width of node name column in the status, top, df and print
command output in order to ensure column alignment when longer node names
(such as "worker-1-10") are in use.  Decreased width of some columns in
the top command output in order to keep the line length less than 80.
Increased length of the divider line in the capstats output.