Fix broctl top on OS X Mavericks
The top command output changed in OS X Mavericks, which was breaking
the broctl top command (the column numbers that the top helper script
needs to use are different in Mavericks).  Also noticed that the grep
command used by the broctl top helper script was not completely
filtering out the first sample output of top (because there are
multiple launchd processes).  Fixed both issues (should also work
on older versions of OS X).