Fix test diff canonifiers that read files on cmdline
The test diff canonifiers that used commands (such as "sort" or "head")
that read input from either stdin or from a command-line filename would
cause btest-diff to never report diffs, so that some failed tests were
not reported as failing.  This is due to btest-diff not working as
documented (rather than just reading from stdin, it also passes the
original filename as a command-line argument, which results in the
canonified baseline always matching the current canonified test
output file for the affected test diff canonifiers).

Fixed by changing the affected test diff canonifiers so they are
immune to this problem.