Make sure logs are archived after broctl kills Bro
During a "broctl stop", if a Bro node doesn't terminate within
a set time interval (defined by the broctl option StopTimeout),
then broctl kills that Bro node with SIGKILL.  In this scenario,
it is unlikely that Bro would have had a chance to execute
"archive-log", so any unarchived log files would be removed as part
of the normal cleanup (via "post-terminate") that happens when
"broctl stop" is issued.

Fixed by adding a new command-line option to "post-terminate"
indicating that the Bro node was killed and "archive-log"
needs to be run in order to archive any remaining log files.

Addresses BIT-1126