Improve start/stop command output for crashed nodes
When stopping a node that was crashed, moved the informational
message to be displayed before doing the crash report because it can
take a while to complete (and not seeing any output during that time
is not helpful to the user).  When starting a node that was crashed,
changed the message to indicate that the node was crashed (doing
the crash report can take a while, so the new message helps the
user understand what's happening).

Also added a test for stopping a node that crashes during shutdown.