Fix start helper script to return nonzero on error
Fixed the start helper script to return nonzero when it gets a PID
value of -1 from the run-bro script, which indicates that an error
occurred before attempting to execute Bro.  Previously when this
happened, the start helper would return 0 (indicating success), and
broctl would show a "bro still initializing" message, which doesn't
make sense.