Fix use of memlimit broctl option
Fixed a typo in the run-bro script that was causing the memlimit option
to be ignored (the ulimit commands in run-bro were not setting anything
because LIMIT was an empty string).

Added another ulimit command to show the new limits for debugging
purposes.

Also added a test to verify that memlimit is not being ignored.