Better error handling for "check" and "scripts" commands
Improved the "check" and "scripts" commands to report a single error
message (instead of multiple confusing messages) when a user forgets to
first do an "install".  Updated documentation of "check" command to
remove a statement about check being the only command that works without
first doing an install.

Also fixed exit status to be non-zero when the check or scripts commands
fail to create or remove tmp directory.