Change the broctl exec command to run only once per host
Changed the broctl exec command to run only once per host instead of
running once for each bro instance.  Changed the output format slightly
by including both the node and host names in order to avoid confusion.

Changed the hosts() function so that it preserves the order of the returned
node list as it was sorted by the nodes() function.  This keeps the output
ordering consistent regardless of whether a command uses nodes() or hosts().