Code reorganization for the getDf function
Reorganized the code for the getDf function so that it no longer outputs
anything.  Instead, the results (along with any errors) are stored in a
list that the function returns so that the caller can decide what to do.
This also avoids reporting the same error message multiple times for the
same host.