Change hosts() function in the plugin API
Changed the hosts() function to return a list of nodes instead of just
hostnames (if needed, the hostnames can easily be extracted) so that the
function is more useful.  For example, now it can be used to prepare
a list of nodes, one per host, in order to execute some command once
per host.