Integrate ssh_runner into broctl
Integrate ssh_runner into broctl by replacing some functions in execute.py
with a wrapper class.  Modified ssh_runner so it knows which addresses
are local in order to avoid trying to establish ssh connection to
localhost.  Modified ssh_runner so that the user can choose to have some
commands interpreted by the shell (this is needed when shell metacharacters
or env. vars. are being used).  Simplified the helper scripts to not
echo the exit code (ssh_runner gets the exit status directly).

Updated some test baselines, and temporarily turned off some failing tests.
Improved the diff-remove-abspath canonifier to stop removing pathname
when a space is encountered so that other cmd-line args are not removed.