broctl cron can now start/stop nodes to match expected state
BroControl now records the expected running state of a node at the
beginning of each start or stop.  This expected state might differ from
the actual state because the start or stop could fail for one or more
nodes.  BroControl cron now compares the two states and starts or stops the
affected nodes as needed.  Previously, broctl cron could never stop a node,
and could only start nodes that were in the "crashed" state.

Addresses BIT-1228