Improve the run-bro and start shell scripts
run-bro:
Moved signal handler to beginning of script to reduce the window of time when
the script could terminate without writing a ".pid" file.  Removed some
unneeded code.

start:
Do not echo a pid of -1 (indicates an error), to avoid having that appear
in the error message printed by broctl.

For both scripts, added more error checks to better communicate failures
to the user.