Improve a few shell scripts
Fixed a minor bug in the summarize-connections script where the LIMIT variable
was always set to an empty string (regardless of whether memlimit was
defined or not), which was causing the ulimit commands to have no effect.
Improved comments in make-archive-name, send-mail, and summarize-connections
scripts.  Improved error detection and reporting for the send-mail script.