Improve stats.log error handling
A user recently reported an issue where somehow spool/stats.log became
corrupt so that stats-to-csv would fail, which would result in an email
containing an error message being sent every time broctl cron ran (because
the failure of stats-to-csv was preventing stats.log from being removed).
Since spool/stats.log was being appended to the copy in the stats dir before
stats-to-csv ran, the copy in the stats dir was growing at a very fast
rate (because the spool/stats.log was never being removed).

Fixed by rearranging some code so that even if stats-to-csv fails
broctl cron will remove spool/stats.log (after appending it to the
copy in the stats dir).  Also improved error messages by providing more
info to the user about the cause of the problem.

Also added a couple more error checks for operations involving stats.log.