Rename broctl.dat to state.db
Renamed the broctl state file because the file format is completely
different compared to previous versions of broctl.  This is to avoid a
problem where someone using an older version of broctl does an in-place
upgrade and broctl reports an error because the file is no longer in
the expected format.