Add new broctl option to make status command run faster
Added a new broctl option "StatusCmdShowAll" to enable users to easily
change the default behavior of the status command so that it runs faster.
Since the status command spends most of its time waiting for broccoli to
gather the peer counts, by omitting that column the status command runs
roughly three to five times faster.