Allow using shorter interface name with PF_RING+DNA
Allow users to specify module name of "dnacl" instead of "dnacluster"
when specifying the interface name in the node config when using
PF_RING+DNA with pfdnacluster_master.  A user on the mailing list
reported a problem when trying to use the longer name with more than 10
app instances, and the suggested solution was to use a shorter
interface name.