Fix a race condition that results in data loss
In the ssh_runner muxer code, when a file descriptor reaches EOF,
we need to immediately remove it from the list of file descriptors.
If it is not removed immediately, then it is sometimes returned from
select() two consecutive times.  If this happens, then the "waiting"
counter might be decremented to zero before all of the data has been
read on the other file descriptor associated with that process.