Python 3 compatibility fixes
Fixed some python 3 issues in ssh_runner.  Also, a small optimization
was made so that the get_muxer function is called only during
initialization of the SSHMaster class instead of every time we send
a command.

Removed a sorted() in the main broctl script that didn't work in
python 3 due to unorderable types.  It was redundant anyway, because
the results are already sorted by the cmdresult module.