Simplify and cleanup code that loads plugins
Removed unused code (libdirinternal is already in sys.path, and a few
functions don't need to return a value as this isn't used anywhere).

Added catching of exceptions raised by the init() method of plugins
so that a badly written plugin doesn't terminate broctl.

Changed a few error messages to warnings because these don't
terminate broctl.