Fix to prevent the broctl stop command from hanging
The broctl stop command runs the post-terminate script, and that script
runs some commands in a subshell in the background so that broctl stop
doesn't have to wait for it to finish.  However, the ssh_runner muxer
code was waiting (via select on stdout and stderr) for the subshell
to finish (which can take minutes) because the subshell inherited
stdout and stderr.  Fix is to redirect these to /dev/null.