Use daemon threads only for remote hosts
Changed the ssh_runner code to use daemon threads only for hosts that
are not local.  For the local host, the worker thread should never hang
so it doesn't need to be a daemon thread.  This change seems to fix
the problem of sometimes seeing a python stack trace upon exit from
broctl caused by daemon threads still running when the python interpreter
shuts down.