Set correct python path in python scripts
Fixed a problem where the python and shell scripts created by
the InstallShellScript macro were not being installed.  This prevented
broctl from working whenever the "--with-python" configure
option was needed to specify a python compatible with broctl.
Fixed a similar problem with the python path in ssh_runner.py (it is
now configured automatically by cmake).