Fix the config change warnings on Python 3
The hash() function on Python 3 does not give same results as Python 2,
and doesn't even give same result between two runs of the Python
interpreter.  Because of this, broctl on Python 3 would always
complain that the config was changed, even when it didn't change.
Fixed by using CRC32, which is consistent on all versions of Python.