Improve error messages involving the state database file
When the python sqlite3 module raises an exception, it provides an error
message that doesn't contain much useful information for users, so broctl
needs to supply additional context to the error message.
For example, when someone installs Bro as the superuser, then
tries to run broctl as an ordinary user the database cannot be created
due to the directory being owned by the root user.

Addresses BIT-1397