Don't use daemon threads in ssh_runner
The behavior of daemon threads in Python seems to be somewhat
version-specific.  Note that not using them introduces a short
(10 secs or less) delay when exiting broctl if a connection to a host
goes down while a worker thread is checking if that host is alive,
however this should be very rare (and can never happen if broctl
is not being run in interactive mode).