Capture output of background post-terminate script
If a spool/tmp/post-terminate directory doesn't get removed, then there
will be a new file "post-terminate.out" which might contain error
messages from the part of the post-terminate script that was run
in the background (these messages would most likely be from the
archive-log script that was executed by post-terminate).
This might be helpful when trying to diagnose a failure to archive
log files.