Rename logs when archive-log is run with "-c" option
When the post-terminate script is dealing with a crashed Bro node,
it runs archive-log with the "-c" flag to indicate that the logs
should be archived but the originals should not be removed.
As a result, a user may be confused when seeing log files in
a spool/tmp/post-terminate directory because it is not clear
which ones were successfully archived.

To avoid confusion, now the original log files will be renamed
(e.g., conn.log --> conn.log_archived) after being successfully
archived.