Keep track of both loaded plugins and active plugins
If a plugin's init() returns True, then the plugin is considered active.
Added an attribute to the Plugin class to keep track of which plugins
are currently active.

A loaded plugin is any plugin found that passes some basic sanity checks.
Changed the list of all plugins to include all loaded plugins instead
of just the active plugins.  This is needed to support config
reloading, as the active status of a plugin may change when a config
change happens.

Also changed some warning messages so that the terminology is
more clear.