Add ability for broctl to reload its configuration
Added the ability for broctl to reload its configuration (both
broctl.cfg and node.cfg).  Each time a user runs a broctl command
a check of the broctl config files is performed to determine if the
config has changed since the last time the config was loaded.  If a
change is detected, then before doing anything else, broctl will either
reload the config (for the "deploy" command) or warn the user to run
the "deploy" command (for any other command).