Send email if post-terminate fails to archive logs
Changed the post-terminate script to run archive-log serially instead
of multiple instances simultaneously in the background, in order to
reduce the load on the system.  Also changed post-terminate to send an
email if it fails to archive one or more logs (the email contains the
name of the specific spool/tmp/post-terminate directory so that the user
can check for error messages or manually archive any logs there).