Improve a couple of diagnostic messages
When a user runs "cleanup --all" now there is a message explaining
what it's doing (previously, there was no indication that it does
anything different than just "cleanup").

Added some useful context to an archive-log error message (this can be
useful if post-terminate tries to archive a file with a strange name,
such as ".crash-diag.log").