Improve check for dangling Bro nodes
Improved the check for dangling Bro nodes by also checking if a node
changed to a different host.  Also added resetting the "expected running"
flag so that broctl cron doesn't try to start such a node.  Also
added clearing the PID in the state store so that we don't keep getting
the same warning.