Fix use of ssh to always use IP address
Changed all uses of ssh to use IP address instead of hostname.
This ensures that the ssh host key lookup always succeeds (assuming that
the user did the manual setup first).  This avoids confusion when a user
uses different combinations of IP addr, hostname, or FQDN when setting
up the ssh host keys and the broctl node.cfg.  For example, if a user
specifies the FQDN in node.cfg, but then manually does an ssh to each
worker machine using just the hostname (not the FQDN) when setting up
the ssh host keys, then "broctl install" would fail because ssh can't
find the host key.  The simplest solution in this case is to have
broctl always use IP addresses with ssh.

Also changed ssh to use BatchMode consistently to avoid a misleading
error message (can't find ssh-askpass) when the rsync fails.